upgraded mpdf library to version 5.1

+++ mpdf/compress.php
@@ -1,7 +1,7 @@
 <?php
 
 $excl = array( 'TABLES', 'LISTS', 'IMAGES-CORE', 
-'IMAGES-WMF', 'TABLES-ADVANCED-BORDERS', 'UNICODE-FONTS', 'HTMLHEADERS-FOOTERS', 'COLUMNS', 'TOC', 'INDEX', 'BOOKMARKS', 'BARCODES', 'FORMS', 'WATERMARK', 'RTL', 'INDIC', 'CJK', 'ANNOTATIONS', 'GRADIENTS', 'BACKGROUND-IMAGES', 'CSS-FLOAT', 'CSS-IMAGE-FLOAT', 'CSS-POSITION', 'CSS-PAGE', 'BORDER-RADIUS', 'HYPHENATION', 'ENCRYPTION', 'DIRECTW', 'PROGRESS-BAR', 'IMPORTS');
+'IMAGES-WMF', 'TABLES-ADVANCED-BORDERS', 'HTMLHEADERS-FOOTERS', 'COLUMNS', 'TOC', 'INDEX', 'BOOKMARKS', 'BARCODES', 'FORMS', 'WATERMARK', 'RTL', 'INDIC', 'ANNOTATIONS', 'BACKGROUNDS', 'CSS-FLOAT', 'CSS-IMAGE-FLOAT', 'CSS-POSITION', 'CSS-PAGE', 'BORDER-RADIUS', 'HYPHENATION', 'ENCRYPTION', 'DIRECTW', 'IMPORTS', 'PROGRESS-BAR');
 
 
 	// *DIRECTW* = Write, WriteText, WriteCell, Text, Shaded_box, AutosizeText
@@ -52,7 +52,7 @@ function checkedAll (frm1) {
 <ul>
 <li>For WMF Images, you must include both IMAGES-CORE and IMAGES-WMF</li>
 <li>JPG, PNG and JPG images are supported with IMAGES-CORE</li>
-<li>IMAGES-CORE are required for BACKGROUND-IMAGES or WATERMARKS to work</li>
+<li>IMAGES-CORE are required for BACKGROUNDS (IMAGES) or WATERMARKS to work</li>
 <li>DIRECTW includes the functions to Write directly to the PDF file e.g. Write, WriteText, WriteCell, Text, Shaded_box, AutosizeText</li>
 </ul>
 </div>
@@ -120,8 +120,15 @@ foreach($l AS $k=>$ln) {
 		$x .= $ln; 
 	}
 }
-
-$check = file_put_contents('mpdf.php', $x);
+// mPDF 5.0
+if (function_exists('file_put_contents')) {
+	$check = file_put_contents('mpdf.php', $x);
+}
+else {
+	$f=fopen('mpdf.php', 'w');
+	$check = fwrite($f, $x);
+	fclose($f);
+}
 if (!$check) { die("ERROR - Could not write to mpdf.php file. Are permissions correctly set?"); }
 echo '<p><b>mPDF file generated successfully!</b></p>';
 echo '<div>mPDF file size '.number_format((strlen($x)/1024)).' kB</div>';
